HistoryPane.js -> HistoryPane.tsx
