+++ src/components/App.tsx
@@ -7,7 +7,7 @@ import ReactDOM from 'react-dom';
 import { INITIAL_ROW_LIMIT } from '../data/constants';
 import { IConfigDatabase } from '../data/models/configModel';
 import Auth from './Auth.js';
-import HistoryPane from './HistoryPane.js';
+import HistoryPane from './HistoryPane';
 import { LeftPane } from './LeftPane';
 import Navigation from './Navigation.js';
 import RightPane from './RightPane.js';
